ad-command-center/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
apscheduler==3.10.4
httpx==0.27.0
openai==1.50.0
python-dotenv==1.0.1
